requirements.txt
httpx==0.28.1
langchain_core==0.3.51
langchain_openai==0.3.12
langgraph==0.3.27
mcp==1.6.0
openai==1.72.0
playwright==1.51.0
pydantic==2.11.3
PyMySQL==1.1.1
python-dotenv==1.1.0
Requests==2.32.3
typing_extensions==4.13.1
